Phase 15-C: extract sinatra-cloudflare-workers gem and split homurabi app
Add path gem with sinatra_opal_patches, JwtAuth/Scheduled/Queue, and
sinatra/cloudflare_workers entry (runtime patches + optional Rack registration).

Move HTTP routes to app/routes/all.rb; thin app/hello.rb uses -I app + require app.
Wire Gemfile and Opal -I paths; add minimal and classic dogfood examples plus
ERB CLI and Rake/wrangler templates.

Made-with: Cursor
