+++ app/routes/api.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+# /api/* routes live in `routes/all.rb`.
