+++ app/routes/demo.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+# Demo / HTML / binding routes live in `routes/all.rb` (single ordered table for Sinatra).
