+++ app/routes/login.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+# /login, /logout, /chat routes live in `routes/all.rb`.
