+++ app/routes/posts.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+# /posts routes live in `routes/all.rb`.
