+++ app/routes/test.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+# /test/* routes live in `routes/all.rb`.
