+++ examples/classic-top-sinatra/Gemfile
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+source 'https://rubygems.org'
+ruby '>= 3.4.0'
+
+gem 'opal', path: '../../vendor/opal-gem'
+gem 'cloudflare-workers-runtime', path: '../../gems/cloudflare-workers-runtime'
+gem 'sinatra-cloudflare-workers', path: '../../gems/sinatra-cloudflare-workers'
