+++ gems/sinatra-cloudflare-workers/lib/sinatra/jwt_auth.rb
