+++ gems/sinatra-cloudflare-workers/lib/sinatra/queue.rb
