+++ gems/sinatra-cloudflare-workers/lib/sinatra/scheduled.rb
