+++ gems/sinatra-cloudflare-workers/templates/Rakefile.example
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+#
+# Example Rake tasks for a Sinatra-on-Cloudflare-Workers project using
+# sinatra-cloudflare-workers + cloudflare-workers-runtime.
+#
+# Copy to your app root as `Rakefile` and adjust paths.
+
+desc 'Precompile ERB views to Ruby (Opal-safe)'
+task 'cloudflare_workers:erb' do
+  sh 'bundle', 'exec', 'cloudflare-workers-erb-compile',
+     '--input', 'views',
+     '--output', 'build/app_templates.rb',
+     '--namespace', 'AppTemplates'
+end
+
+desc 'Compile Opal bundle (same flags as npm run build:opal in homurabi)'
+task 'cloudflare_workers:opal' do
+  sh({ 'OPAL_PREFORK_DISABLE' => '1' }, 'bundle', 'exec', 'opal',
+     '-c', '-E', '--esm', '--no-source-map',
+     '-I', 'gems/cloudflare-workers-runtime/lib',
+     '-I', 'gems/sinatra-cloudflare-workers/lib',
+     '-I', 'lib', '-I', 'vendor', '-I', 'build',
+     '-r', 'opal_patches', '-r', 'cloudflare_workers', '-r', 'app_templates',
+     '-o', ENV.fetch('OPAL_OUTPUT', 'build/app.mjs'),
+     ENV.fetch('OPAL_INPUT', 'app/app.rb'))
+end
+
+desc 'ERB + Opal (adjust to include assets if needed)'
+task 'cloudflare_workers:build' => %w[cloudflare_workers:erb cloudflare_workers:opal]
+
+task default: 'cloudflare_workers:build'
